requirements.txt
Django>=3.1.3,<3.2.0
djangorestframework>=3.12.2,<3.13.0
psycopg2>=2.7,<2.8
Pillow>=5.3.0,<5.4.0
uwsgi>=2.0.18,<2.1.0
boto3>=1.12.0,<1.13.0
django-storages>=1.9.1,<1.10.0
django-cors-headers>=3.5.0,<3.6.0

# flake8>=3.6.0,<3.7.0